Rename stop module to stop_conditions
