+++ curated_transformers/generation/stop_conditions.py
