+++ curated_transformers/tests/generation/test_stop.py
@@ -1,7 +1,7 @@
 import torch
 
 from curated_transformers.generation.state import GeneratorState
-from curated_transformers.generation.stop import (
+from curated_transformers.generation.stop_conditions import (
     EndOfSequenceCondition,
     MaxGeneratedPiecesCondition,
 )
